Make `Query.value` not specific to Query but to the Reindexed Model
Why? To build your query you require the models, if the models
want to define queries using `Query.value` they can't because
you will enter in a cyclic dependency. To avoid that we move
`value` to the top module namespace and alias it in `Query` to
keep it bakwards compatible.